requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
django-debug-toolbar==3.2.2
djangorestframework==3.12.4
Pillow==8.4.0
platformdirs==2.4.0
PyJWT==2.2.0
regex==2021.10.8
sqlparse==0.4.2
